bam code edit unit type

+++ SwDesign2/src/main/java/com/cp/repository/UnitRepository.java
@@ -6,6 +6,6 @@ import com.cp.model.Unit;
 
 
 
-public interface UnitRepository extends CrudRepository<Unit,Long>{
+public interface UnitRepository extends CrudRepository<Unit,Integer>{
 
 }
